requirements.txt
asgiref==3.2.10
Django==3.1.2
django-resized==0.3.11
djangorestframework==3.12.1
Markdown==3.3.1
Pillow==7.2.0
pytz==2020.1
sqlparse==0.4.1
